inventory: Refactor in terms of interfaces
Remove Inventory::updateObject() from the interface and replace it with
Inventory::add() and Inventory::remove() in terms of the Interface
subclasses. With this, we no-longer lose information about how to manage
properties associated with a specific interface, we can add and remove
them as necessary. This is fundamental for the correct operation
of PublishWhenPresentInventoryDecorator which now removes any data
interfaces when the item is marked as absent.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>